Add a Static check
At the moment, consumers of rack-ecg are expected to implement an entire
new check in order to return a specific value. While this encourages
consistency, it also adds barriers when attempting to return consistent
sources of information returned in new ways.

For example, applications built into a Docker/OCI image often choose to
embed a git revision or build number into the image itself, and don't
include the git executable. In this scenario, there is no simple way
to return that revision in a manner consistent with the existing
GitRevision check.

The new Static check accepts a small set of parameters, and uses it to
return an identical result on every invocation. One parameter (name)
allows it to return the results under any desired check key, which
provides a solution to the aforementioned issue.