Deprecate the git_revision check
The git_revision check requires that the git executable is present, and
that the current working directory is a relevant git repository.
Additionally, it runs the git executable to determine the current
revision on every call to Rack::ECG.

Given industry practices around immutable instances, it is unlikely that
the git revision for an application will change at runtime. Therefore,
calling git repeatedly introduces overhead. Additionally, when building
Docker/OCI images, industry practices recommend minimizing extra
dependencies or files - therefore, it is unlikely that either the git
executable, or the git repository data, will be available in the running
image.

Now that the static check exists, there is a migration path to allow
consumers to either:

* migrate to a memoized version of the existing check, or
* expose a result for git_revision via some other means, e.g. recording
  the git revision in an application image at build time, and loading
  that from a file or environment variable.