Make RedisConnection Check Require Redis Instance Parameter
Redis-rb, the Redis gem used by Rails apps, is deprecating a part of its
API which stores a global Redis connection. See
https://github.com/redis/redis-rb/issues/1064 for context around that
deprecation.

To prepare for when this behavior will be removed, we must update this
application so that Redis will be passed in as an instance to be
checked, because it will not be globally available.